+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 UNRELEASED
 
+- Remove ruby concurrency limiter
+  https://gitlab.com/gitlab-org/gitaly/merge_requests/708
 - Introduce src-d/go-git as dependency
   https://gitlab.com/gitlab-org/gitaly/merge_requests/709